pyproject.toml
[project]
name = "qizheng-tool"
version = "0.1.0"
description = "七政四余排盘工具 for 李看看"
requires-python = ">=3.10"
dependencies = [
    "pyswisseph>=2.10.0",
    "skyfield>=1.50",
    "lunarcalendar>=0.1.0",
    "ephem>=4.1.5",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
